Remove typechecking
Not really, but sort of. What we have left is essentially a dynamically
typed language—everything having the JSON type—plus some kind of
polymorphism to catch the most egregious type errors. The point being
that JSON is actually quite dynamic in nature, and to really properly
type check an rq expression we'd need a much stronger type system;
perhaps something akin to [1]. I'd certainly like to implement something
like this *at one point*, but for the time being this feels like another
rabbit-hole, which would delay the release of this even further—possibly
indefinitely. Plus, this being a one-line kind of deal, I'd imagine many
users don't even care *where* errors are reported, because for them
there is very little distinction between the different phases.

[1]: https://arxiv.org/pdf/1306.6032.pdf