expr: Remove desugaring step
Overall, this seems to reduce complexity somewhat. At first I thought
that we'd have a lot more tricky syntactic sugar, so that it's better to
separate the two ASTs. However, this didn't turn out to be true, so it
just seems like a bit of superfluous copying around at this point. We do
loose the 1-1 correspondence between what Expr can represent and what
the parser parses, see [Note Builting] in type/checker.rs, but this
seems like the lesser evil to me.