Add binary operators
By not creating an extra node in the AST, and directly desugaring binary
operators during parsing, the type checking of Builtin in
type/checker.rs is not superfluous anymore! This feels like a win, and
also means that I can go all in with pretty-printing the entered
"program" as an S-expression. The readability of the parser further
deteriorates, but that was to be expected, I guess.