type: Switch to bidirectional type checking
Specifying the standard library in this way is _a tad_ unergonomic
currently (and this extends to type annotation, so it's worth fixing) as
everything has to be a TypVar. I'm thinking the fix is to really only
have the existentials be numbers, and the variables can be strings
instead. This requires some refactoring, for which I'm far too tired
right now, so it's problem for future-me.