util: Add fatten
This probably still has some kinks—and is thus almost entirely
undocumented right now—but the basic functionality seems to be in place,
and I happen to think that the algorithm is kind of cute.

The hope is that this public commit will move me towards actually
finishing and polishing this up a bit.